requirements.txt
Django==5.0.2
django-compat==1.0.15
django-cors-headers==4.7.0
djangorestframework==3.16.1
django-filter>=24,<26
pymysql
django-rest-auth
python-dateutil==2.9.0.post0
requests==2.32.3
cryptography==42.0.8
google-cloud-storage==2.15.0
Pillow==10.4.0

# DEPENDENCIAS PARA RENDER
psycopg2-binary==2.9.11
dj-database-url==3.0.1
whitenoise==6.11.0
gunicorn==23.0.0
brotli==1.2.0
packaging==25.0
click==8.3.1
colorama==0.4.6
h11==0.16.0
uvicorn==0.38.0